apps/api/requirements/base.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.6.0
celery[redis]==5.4.0
redis==5.2.1
psycopg[binary]==3.2.3
drf-spectacular==0.27.2
sentry-sdk==2.19.0
python-dotenv==1.0.1
openpyxl==3.1.5
pandas==2.2.3
XlsxWriter==3.2.0
python-magic==0.4.27
python-slugify==8.0.4
gunicorn==23.0.0
Pillow==11.0.0
xlrd>=2.0.1
python-dateutil>=2.8.2
celery[redis]>=5.3.0
flower>=2.0.0
django-celery-results>=2.5.0
django-celery-beat>=2.5.0
django-storages[boto3]>=1.14.0
boto3>=1.28.0
phonenumbers>=8.13.0
django-redis>=5.4.0

reportlab>=4.0.0
anthropic>=0.36.0
